Split all the BulkAll and ScrollAll related tests into their own files and classes making it easier to read what they are doing
