+++ src/Tests/Tests/Document/Multiple/BulkAll/BulkAllForEachAsyncApiTests.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Reactive.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Elastic.Xunit.XunitPlumbing;
+using FluentAssertions;
+using Tests.Core.ManagedElasticsearch.Clusters;
+using Tests.Framework.Integration;
+
+namespace Tests.Document.Multiple.BulkAll
+{
+	public class BulkAllForEachAsyncApiTests : BulkAllApiTestsBase
+	{
+		public BulkAllForEachAsyncApiTests(IntrusiveOperationCluster cluster, EndpointUsage usage) : base(cluster, usage) { }
+
+		[I] public async Task AwaitBulkAll()
+		{
+			var index = CreateIndexName();
+
+			var size = 1000;
+			var pages = 10;
+			var seenPages = 0;
+			var numberOfDocuments = size * pages;
+			var documents = this.CreateLazyStreamOfDocuments(numberOfDocuments);
+
+			var tokenSource = new CancellationTokenSource();
+			var observableBulk = this.Client.BulkAll(documents, f => f
+					.MaxDegreeOfParallelism(8)
+					.BackOffTime(TimeSpan.FromSeconds(10))
+					.BackOffRetries(2)
+					.Size(size)
+					.RefreshOnCompleted()
+					.Index(index)
+					.BufferToBulk((r, buffer) => r.IndexMany(buffer))
+				, tokenSource.Token);
+
+
+			await observableBulk
+				.ForEachAsync(x => Interlocked.Increment(ref seenPages), tokenSource.Token);
+
+			seenPages.Should().Be(pages);
+			var count = this.Client.Count<SmallObject>(f => f.Index(index));
+			count.Count.Should().Be(numberOfDocuments);
+		}
+	}
+}
